requirements.txt
BeautifulSoup==3.2.1
Django==1.5.5
Fabric==1.6.1
Jinja2==2.7.2
Markdown==2.4
MarkupSafe==0.18
MySQL-python==1.2.5
Pygments==1.6
South==0.8.1
Sphinx==1.1.3
amqp==1.0.13
anyjson==0.3.3
billiard==2.7.3.34
celery==3.0.24
django-appconf==0.6
django-celery==3.0.21
django-compressor==1.3
django-debug-toolbar==1.0.1
django-extensions==1.3.3
django-filter==0.7
django-model-utils==2.0.2
django-rest-swagger==0.1.13
djangorestframework==2.3.12
docutils==0.11
ecdsa==0.10
kombu==2.5.16
numpy==1.8.0
pandas==0.13.1
paramiko==1.12.2
pycrypto==2.6.1
python-dateutil==2.2
pytz==2013.9
six==1.5.2
sqlparse==0.1.11
virtualenv==1.11.4
wsgiref==0.1.2
